deprecated the Chart and StatsChart components and renamed with Old
