--- src/components/charts/Chart/index.ts
@@ -1,3 +1,0 @@
-export * from './types'
-import Chart from './Chart'
-export default Chart
