+++ src/components/charts/OldChart/index.ts
@@ -1,0 +1,3 @@
+export * from './types'
+import OldChart from './OldChart'
+export default OldChart
