+++ src/components/charts/OldStatChart/OldStatsChartStyles.ts
