+++ src/components/charts/OldStatChart/index.ts
@@ -1,0 +1,3 @@
+export * from './types'
+import OldStatsChart from './OldStatsChart'
+export default OldStatsChart
