+++ src/components/charts/OldStatChart/types.ts
