--- src/components/charts/StatChart/index.ts
@@ -1,3 +1,0 @@
-export * from './types'
-import StatsChart from './StatsChart'
-export default StatsChart
